scripts/openapi-gen/requirements.txt
pyyaml>=6.0
requests>=2.28.0
protobuf>=4.21.0
grpcio-tools>=1.48.0
click>=8.0.0
jinja2>=3.0.0